abstract out WalSet
