+++ db/version_set.h
@@ -1174,7 +1174,7 @@ class VersionSet {
     AppendVersion(cfd, version);
   }
 
-  const std::map<WalNumber, WalMetadata>& GetWals() const { return wals_; }
+  const WalSet& GetWalSet() const { return wals_; }
 
  protected:
   using VersionBuilderMap =
@@ -1250,15 +1250,11 @@ class VersionSet {
   Status VerifyFileMetadata(const std::string& fpath,
                             const FileMetaData& meta) const;
 
-  // The WALs of the DB.
-  // When a WAL is created or closed, a VersionEdit is applied to VersionSet.
-  // WALs are archived or deleted asynchronously, so even if a WAL has been
-  // archived or deleted, it might still exist in wals_.
-  // If we can determine that a WAL is no longer used based on the log numbers
-  // kept in MANIFEST, it'll be deleted from wals_ in the following situations:
-  // 1. when writing the first snapshot during creating a new MANIFEST;
-  // 2. when recovering from MANIFEST.
-  std::map<WalNumber, WalMetadata> wals_;
+  // WAL with log number less than this returned value can be ignored during
+  // WAL recovery.
+  WalNumber MinLogNumberToKeep() const;
+
+  WalSet wals_;
 
   std::unique_ptr<ColumnFamilySet> column_family_set_;
   Env* const env_;
